# Add start time on perf trace because it seems some steps haven't been recorded.